SmoothAnim Move #CHANGE: Moved smooth animation default resources to the SmoothAnimation sample. #CHANGE: Moved VertexC1P3T4A1 vertex format to the SmoothAnimation sample.
