+++ Samples/SmoothAnimation/Content/Data/SmoothAnimation/Visuals/SmoothAnim.ShaderProgram.res
@@ -1,0 +1,10 @@
+﻿<root dataType="Struct" type="Duality.Resources.ShaderProgram" id="129723834">
+  <assetInfo />
+  <frag dataType="Struct" type="Duality.ContentRef`1[[Duality.Resources.FragmentShader]]">
+    <contentPath dataType="String">Data\SmoothAnimation\Visuals\SmoothAnim.FragmentShader.res</contentPath>
+  </frag>
+  <vert dataType="Struct" type="Duality.ContentRef`1[[Duality.Resources.VertexShader]]">
+    <contentPath dataType="String">Data\SmoothAnimation\Visuals\SmoothAnim.VertexShader.res</contentPath>
+  </vert>
+</root>
+<!-- XmlFormatterBase Document Separator -->
