+++ Samples/SmoothAnimation/SmoothAnimation.csproj
@@ -49,6 +49,7 @@
     <Compile Include="SmoothAnimSpriteRenderer.cs" />
     <Compile Include="CorePlugin.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="VertexC1P3T4A1.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\Source\Core\Primitives\DualityPrimitives.csproj">
